#3 Move tmp to localcopy
