+++ src/senzing/g2engine_grpc.py
@@ -11,9 +11,9 @@ from typing import Any, Tuple
 import grpc  # type: ignore
 
 from .g2helpers import as_str
+from .localcopy.g2engine_abstract import G2EngineAbstract
+from .localcopy.g2engineflags import G2EngineFlags
 from .pb2_grpc import g2engine_pb2, g2engine_pb2_grpc
-from .tmp.g2engine_abstract import G2EngineAbstract
-from .tmp.g2engineflags import G2EngineFlags
 
 # Metadata
