+++ src/senzing/localcopy/g2config_abstract.py
