+++ src/senzing/localcopy/g2configmgr_abstract.py
