+++ src/senzing/localcopy/g2diagnostic_abstract.py
