+++ src/senzing/localcopy/g2engineflags.py
