+++ src/senzing/localcopy/g2hasher_abstract.py
