+++ src/senzing/localcopy/g2product_abstract.py
